requirements.txt
##### Without version #####
qiskit
quimb
cotengra
sparse
mqt.ddsim
optuna
kahypar
networkx
matplotlib
numpy


##### With version    #####


